api/requirements.txt
flask==3.1.0
flask-cors==5.0.1
snowflake-connector-python==3.12.3
requests==2.32.3
